% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignMS2s.R
\name{alignMS2s}
\alias{alignMS2s}
\title{alignMS2s}
\usage{
alignMS2s(sweeperObj, ppm = 5, betweenSampThresh = 0.15,
  poisThresh = 0.05, cores = 1, mzDiff = 0.5)
}
\arguments{
\item{sweeperObj}{- sweeper object containing ms2 data for each matched
feature per sample}

\item{ppm}{- ppm error to check for similarities between parent ions across
samples.}

\item{betweenSampThresh}{- minimum p probability value required to retain a
peak during across sample comparison.}

\item{poisThresh}{- minimum p value required to retaion a peak during
modeling of occurance frequencies.}

\item{cores}{- number of cores running the algorithm.}

\item{mzDiff}{- absolute mass difference required to group common peaks of
across distinct ms2 scans.}
}
\value{
the sweeper object with the pureMS2 slot filled.
}
\description{
This function is designed to align MS2 spectra from a common
parent ion. The input consists of the sweeperObj and out output is the
same object with the pureMS2 slot filled.
}
