% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check13c.R
\name{check13c}
\alias{check13c}
\title{check13c}
\usage{
check13c(windowData, ppm)
}
\arguments{
\item{windowData}{- the mz intensity table making up the observations of a
given mass spectra.}

\item{ppm}{- error window used to identify 13C peaks}
}
\value{
The windowData table with an additional column used to indicate
if the data has a 13C peak that could be contamination the ms2 scan.
}
\description{
This function checks if the scan with the parent ion has a
13C feature. If it does, it remove the 13C feature from the scan window and
combines the purity. The new table tells us if we need to fix the ms2 should
be corrected for contamination.
}
