% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweeperObj-class.R
\name{createSweeper}
\alias{createSweeper}
\title{createSweeper}
\usage{
createSweeper(features, ms2Path)
}
\arguments{
\item{features}{- two column data.frame representing m/z and rt values of
individual features to be screened for MS2 spectra.}

\item{ms2Path}{- path to raw ms2 data.}
}
\description{
This function will create a sweeperObj used to extract ms2s.
}
