% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFragments.R
\name{filterFragments}
\alias{filterFragments}
\title{filterFragments}
\usage{
filterFragments(curMz, curRt, sampleData, ppm, rtError)
}
\arguments{
\item{curMz}{- current m/z value of feature searched within the data}

\item{curRt}{- current rt value of feature searched within the data}

\item{sampleData}{- current sample being screened for ms2 matches}

\item{ppm}{- mass error match between sample m/zs and entered curMz}

\item{rtError}{- rt error between sample feature's rt and curRt}
}
\value{
a vector of matching indicies.
}
\description{
This function is designed to find mz and rt matches between the
current feature being investigated and a sample. It shoud return indicies of
features that are within the allowed error.
}
