% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMS2s.R
\name{filterMS2s}
\alias{filterMS2s}
\title{filterMS2s}
\usage{
filterMS2s(sweeperObj, minScanCount = 0, minIntensity = 10)
}
\arguments{
\item{sweeperObj}{- Sweeper Object containing MS2 spectra.}

\item{minScanCount}{- Minimum number of scan containing a peak.}

\item{minIntensity}{- Minimum intensity value required to retain a peak.}
}
\value{
returns the sweeperObj with the filttered ms2 spectra by the
user supplied heuristic values.
}
\description{
This function is designed to apply user input filters to filter
ms2 peaks of the data. Currently this funtion filters by scan count and by
minimum intensity value.
}
