% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix13C.R
\name{fix13C}
\alias{fix13C}
\title{fix13C}
\usage{
fix13C(ms2Table)
}
\arguments{
\item{ms2Table}{- ms2 spectra table being checked for 13C peaks.}
}
\value{
A ms2 spectra table corrected for 13C peaks.
}
\description{
This function is designed to remove peaks that may be from 13C
peak within the MS2 scan
}
