% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweeperObj-class.R
\name{getFreatures}
\alias{getFreatures}
\title{get_set}
\usage{
getFreatures(sweeperObj)
}
\description{
This file contains the getter and setter functions used to
access slots from sweeperObj object.
}
