% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestMs2s.R
\name{harvestMS2}
\alias{harvestMS2}
\title{harvestMS2}
\usage{
harvestMS2(sweeperObj, ppm = 5, rtError = 15, isoWindow = 2,
  scanInterval = 7, cores = 3, clearData = TRUE, mzDiff = 0.5)
}
\arguments{
\item{sweeperObj}{- object containing all MS2 data that will be mined.}

\item{ppm}{- Error window for features.}

\item{rtError}{- Time window to check for matches in rt betwee parent ions
from XCMS and those from the MS2 fragment table.}

\item{isoWindow}{- isolation window for MS2s}

\item{scanInterval}{- Dynamic exclusion time. Usefull when matching MS2s}

\item{cores}{- number of cores to use during harvest function.}

\item{clearData}{- Boolean. Deletes raw data from sweepersObj. Good to
reduce amount of memory taken up by sweeperObj.}

\item{mzDiff}{- mz error window used to group peaks from spectra together.
that belong to the same features but come in different scans.}
}
\value{
sweeperObj with slots for parents and daughters filled.
}
\description{
This function connects all MS2 scans belonging to a single feature
within a single data sample. It returns the sweeperObj with a slot for the
parent ions and a slot for the merged MS2s filled up. It also does a 13C
correction of the MS2s if a 13C peak is detected within the parent ion's
scan.
}
