% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweeperObj-class.R
\docType{methods}
\name{initialize,sweeperObj-method}
\alias{initialize,sweeperObj-method}
\title{Update an \code{\linkS4class{sweeperObj}} object}
\usage{
\S4method{initialize}{sweeperObj}(.Object, features, ms2Path)
}
\arguments{
\item{.Object}{- the \code{\linkS4class{sweeperObj}} object to update.}

\item{features}{- a two column data.frame containing features to be
screened for MS2s.}

\item{ms2Path}{- file path the raw MS2 data.}
}
\value{
An updated \code{\linkS4class{sweeperObj}} containing all data from
the input object.
}
\description{
This method updates an \code{\linkS4class{sweeperObj}}
    object to the latest definition.
}
\author{
Craig McLean
}
