% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateIntensity.R
\name{interpolateIntensity}
\alias{interpolateIntensity}
\title{interpolateIntensity}
\usage{
interpolateIntensity(ms2Matches, purityTable, curMs2)
}
\description{
This funciton is meant to interpolate the proper intensity value for the MS2
between the two adjacent peaks. This is done in attempt to correct the spectra from convolution
due to noise of features.
}
