% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Info.R
\name{ms2Info}
\alias{ms2Info}
\title{ms2Info}
\usage{
ms2Info(MSn_dt)
}
\arguments{
\item{MSn_dt}{- list of tables containing all fragments observed}
}
\description{
This function is used to subset entered peak table for all fragments into only values
related to MS2 observations.
}
