% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentPurity.R
\name{parentPurity}
\alias{parentPurity}
\title{parentPurity}
\usage{
parentPurity(sampleData, curMz, parentMatches, allMasses, allIntensities,
  scanTimes, ppm, isoWindow)
}
\arguments{
\item{sampleData}{- current sample of MS2 data being looked at.}

\item{curMz}{- m/z being queried within data.}

\item{parentMatches}{- index of sampleData that contain matches to curMz}

\item{allMasses}{- all mz values from sample raw data}

\item{allIntensities}{- all intensity values from sample raw data}

\item{scanTimes}{- all scans (rts) from sample raw data}

\item{ppm}{- mass error}

\item{isoWindow}{- a measure in dalton of the isolation error from the instrument}
}
\value{
List of parent m/z matches over time for each feature along with their purity.
}
\description{
This is a function used to determine the purity of each parent
ion in making the ms2 spectra. The amuWindow defines the range of masses
that can enter the collision cell. Ideally, this will be used to help correct
patterns and deconvolute the ms2 spectra.
}
