% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pureDaughters.R
\name{pureDaughters}
\alias{pureDaughters}
\title{pureDaughters}
\usage{
pureDaughters(ms2Matches, mzDiff)
}
\arguments{
\item{ms2Matches}{- a list of all ms2 matches that belong to features
within the error of the input curMz and curRt values.}

\item{mzDiff}{- absolute mz error required match spectra across scans.}
}
\value{
a list of tables making up the linked scans within a sample.
}
\description{
This function works the same as the pureMS2 function, but is
slightly different for matching spectra across samples. This only means
that it has a few more error messages.
}
