% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pureMS2.R
\name{pureMS2}
\alias{pureMS2}
\title{pureMs2}
\usage{
pureMS2(ms2Matches, mzDiff)
}
\arguments{
\item{ms2Matches}{- a list of all ms2 matches that belong to features
within the error of the input curMz and curRt values.}

\item{mzDiff}{- absolute mz error required match spectra across scans.}
}
\value{
a list of tables making up the linked scans within a sample.
}
\description{
This function is designed to return a high confidence MS2 table. It mataches features
that are found within two tables.
}
