% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retreiveSpectra.R
\name{retreiveSpectra}
\alias{retreiveSpectra}
\title{retreiveSpectra}
\usage{
retreiveSpectra(sweeperObj, outputPath)
}
\arguments{
\item{sweeperObj}{- sweeper object that has been used throughout the
algorithm.}

\item{outputPath}{- path to store output files to run on metfrag.}
}
\description{
This function is desgined to output MS2 spectra gathered from
the sweeper algorithm in feature (mz/rt) specfic tables.
}
