% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMs2s.R
\name{summarizeMs2s}
\alias{summarizeMs2s}
\title{summarizeMs2s}
\usage{
summarizeMs2s(sweeperObj)
}
\arguments{
\item{sweeperObj}{The ms2 sweeper object containing all relevant info.}
}
\description{
This function is designed to return some stats on the quality
of MS2. The idea here is to get a feel for what might be an appropriate
filter to clean up the data.
}
