% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestMs2s.R
\name{harvestMS2}
\alias{harvestMS2}
\title{harvestMS2}
\usage{
harvestMS2(sweeperObj, ppm = 5, rtError = 15, isoWindow = 2,
  scanInterval = 7, cores = 3, clearData = TRUE, mzDiff = 0.5)
}
\description{
This function connects all MS2 scans belonging to a single feature
within a single data sample. It returns the sweeperObj with a slot for the
parent ions and a slot for the merged MS2s filled up. It also does a 13C
correction of the MS2s if a 13C peak is detected within the parent ion's
scan.

 @param sweeperObj - object containing all MS2 data that will be mined.
 @param rtError - Time window to check for matches in rt betwee parent ions
 from XCMS and those from the MS2 fragment table.
 @param ppm - Error window for features.
 @param isoWindow - isolation window for MS2s
 @param scanInterval - Dynamic exclusion time. Usefull when matching MS2s
 @param clearData - Boolean. Deletes raw data from sweepersObj. Good to
 reduce amount of memory taken up by sweeperObj.
 @param mzDiff - mz error window used to group peaks from spectra together.
 that belong to the same features but come in different scans.
 @param cores - number of cores to use during harvest function.

 @importFrom foreach "%dopar%"

 @return sweeperObj with slots for parents and daughters filled.
 @export
}
