% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateIntensity.R
\name{interpolateIntensity}
\alias{interpolateIntensity}
\title{interpolateIntensity}
\usage{
interpolateIntensity(purityTable)
}
\arguments{
\item{purityTable}{- Table of parent ion features observed over time.}
}
\value{
coefficients from running a regression on the parent ions over
the time window where the peak is formed.
}
\description{
This funciton is meant to interpolate the proper intensity value for the MS2
between the two adjacent peaks. This is done in attempt to correct the spectra from convolution
due to noise of features.
}
