% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweeperObj-class.R
\docType{class}
\name{sweeperObj-class}
\alias{sweeperObj-class}
\title{sweeperObj}
\description{
This object is a generic object designed to run the different
functions of the ms2sweeper package. The slots represent content or data
that the package uses throughout the different functions.
}
\section{Slots}{

\describe{
\item{\code{features}}{- A table containing an m/z and rt column used to select
specific features from ms2 data.}

\item{\code{ms2Path}}{- A string vector of paths to the ms2 raw data files.}

\item{\code{ms2Data}}{- The read in ms2 data.}

\item{\code{ms2Harvest}}{- Harvested ms2 data for all inputed parent ions from
features.}

\item{\code{ms2Pure}}{- list of purified ms2s based on the cleaning criterion.}
}}

